Finished HW # 7
I added the needed/mentioned files from the homework. To run Newman 3.3 it will have to be on a computer with the right libraries available for the animation. Otherwise everything is together. I did an extra object oriented conversion on accident, but it was good practice so I left it in.